federated-tabular-health/requirements.txt
flwr>=1.8.0
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3
torch>=2.1
matplotlib>=3.7
pyyaml>=6.0
pytest>=7.4
